+++ util/ca.go
@@ -29,3 +29,15 @@ func IsSelfSigned(c *x509.Certificate) bool {
 func IsSubscriberCert(c *x509.Certificate) bool {
 	return !IsCACert(c) && !IsSelfSigned(c)
 }
+
+func IsTestableBRCertificate(cert *x509.Certificate) bool {
+	if len(cert.ExtKeyUsage) == 0 {
+		return true
+	}
+	for _, eku := range cert.ExtKeyUsage {
+		if eku == x509.ExtKeyUsageAny || eku == x509.ExtKeyUsageServerAuth {
+			return true
+		}
+	}
+	return false
+}
